First draft of Riker reorganization.
I have created this document to get buy-in for forming a core team for the Riker project. I believe it is important to communicate the vision and set basic direction. It's critical to get basic alignment before discussing to-dos. After agreeing on these basics we can discuss concrete action points, such as where/how to communicate, near-term roadmap, code review & merge process, etc.

You are welcome to propose changes to this document directly or comment below.